+++ src/Route/Post.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Router\Route;
+
+use Yiisoft\Http\Method;
+
+final class Post extends MethodRoute
+{
+    protected const METHOD = Method::POST;
+}